# gen/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "contact": {
            "email": "support@codespace.dev",
            "name": "Codespace Operator Support",
            "url": "https://github.com/codespace-operator/codespace-operator"
        },
        "description": "REST API for managing Codespace sessions with RBAC and multi-provider authentication",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "https://github.com/codespace-operator/codespace-operator/blob/main/TERMS.md",
        "title": "Codespace Operator API",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/introspect/server": {
            "get": {
                "description": "Get server and cluster information",
                "parameters": [
                    {
                        "description": "Whether to discover namespaces (0 or 1)",
                        "in": "query",
                        "name": "discover",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ServerIntrospectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Server introspection",
                "tags": [
                    "user"
                ]
            }
        },
...
        "/api/v1/server/sessions": {
            "get": {
                "description": "Get a list of codespace sessions, optionally across all namespaces",
                "operationId": "listSessions",
                "parameters": [
                    {
                        "description": "Target namespace",
                        "in": "query",
                        "name": "namespace",
                        "schema": {
                            "type": "string",
                            "default": "default"
                        }
                    },
                    {
                        "description": "List sessions across all namespaces",
                        "in": "query",
                        "name": "all",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.SessionListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "List sessions",
                "tags": [
                    "sessions"
                ]
            },
            "post": {
                "description": "Create a new codespace session",
                "operationId": "createSession",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/internal_server.SessionCreateRequest"
                            }
                        }
                    },
                    "description": "Session creation request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.Session"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Create session",
                "tags": [
                    "sessions"
                ]
            }
        },
...
    },
...
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer token authentication (JWT)",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            },
            "CookieAuth": {
                "description": "Session cookie authentication",
                "in": "cookie",
                "name": "codespace_session",
                "type": "apiKey"
            }
        },
        "schemas": {
            "common.AnchorMeta": {
                "properties": {
                    "name": {
                        "description": "release/app/deployment name (sanitized when used as label)",
                        "type": "string"
                    },
                    "namespace": {
                        "description": "namespace where the manager runs",
                        "type": "string"
                    },
                    "type": {
                        "description": "argo|helm|deployment|statefulset|daemonset|cronjob|job|pod|namespace|unresolved",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.AuthSpec": {
                "properties": {
                    "mode": {
                        "description": "+kubebuilder:validation:Enum=oauth2proxy;none\n+kubebuilder:default=none",
                        "type": "string"
                    },
                    "oidc": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.OIDCRef"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.NetSpec": {
                "properties": {
                    "annotations": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object"
                    },
                    "host": {
                        "type": "string"
                    },
                    "tlsSecretName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.OIDCRef": {
                "properties": {
                    "clientIDSecret": {
                        "type": "string"
                    },
                    "clientSecretRef": {
                        "type": "string"
                    },
                    "issuerURL": {
                        "description": "+kubebuilder:validation:Pattern=`^https?://`",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.PVCSpec": {
                "properties": {
                    "mountPath": {
                        "description": "+kubebuilder:validation:MinLength=1",
                        "type": "string"
                    },
                    "size": {
                        "description": "+kubebuilder:validation:Pattern=`^\\d+(Gi|Mi)$`",
                        "type": "string"
                    },
                    "storageClassName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.ProfileSpec": {
                "properties": {
                    "cmd": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "ide": {
                        "description": "+kubebuilder:validation:Enum=jupyterlab;vscode;rstudio;custom",
                        "type": "string"
                    },
                    "image": {
                        "description": "+kubebuilder:validation:MinLength=1",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.Session": {
                "properties": {
                    "apiVersion": {
                        "description": "APIVersion defines the versioned schema of this representation of an object.\nServers should convert recognized schemas to the latest internal value, and\nmay reject unrecognized values.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources\n+optional",
                        "type": "string"
                    },
                    "kind": {
                        "description": "Kind is a string value representing the REST resource this object represents.\nServers may infer this from the endpoint the client submits requests to.\nCannot be updated.\nIn CamelCase.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds\n+optional",
                        "type": "string"
                    },
                    "metadata": {
                        "$ref": "#/components/schemas/v1.ObjectMeta"
                    },
                    "spec": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.SessionSpec"
                    },
                    "status": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.SessionStatus"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.SessionSpec": {
                "properties": {
                    "auth": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.AuthSpec"
                    },
                    "home": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.PVCSpec"
                    },
                    "networking": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.NetSpec"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.ProfileSpec"
                    },
                    "replicas": {
                        "type": "integer"
                    },
                    "scratch": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.PVCSpec"
                    }
                },
                "type": "object"
            },
            "github_com_codespace-operator_codespace-operator_api_v1.SessionStatus": {
                "properties": {
                    "phase": {
                        "description": "Pending | Ready | Error",
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_server.CasbinPermissions": {
                "properties": {
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.NamespacePermissions"
                    }
                },
                "type": "object"
            },
            "internal_server.ClusterInfo": {
                "properties": {
                    "casbin": {
                        "$ref": "#/components/schemas/internal_server.CasbinPermissions"
                    },
                    "serverServiceAccount": {
                        "$ref": "#/components/schemas/internal_server.ServiceAccountInfo"
                    }
                },
                "type": "object"
            },
...
            "internal_server.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_server.NamespacePermissions": {
                "properties": {
                    "list": {
                        "type": "boolean"
                    },
                    "watch": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "internal_server.ServerIntrospectionResponse": {
                "properties": {
                    "capabilities": {
                        "$ref": "#/components/schemas/internal_server.SystemCapabilities"
                    },
                    "cluster": {
                        "$ref": "#/components/schemas/internal_server.ClusterInfo"
                    },
                    "instanceID": {
                        "type": "string"
                    },
                    "manager": {
                        "$ref": "#/components/schemas/common.AnchorMeta"
                    },
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.ServerNamespaceInfo"
                    },
                    "version": {
                        "$ref": "#/components/schemas/internal_server.ServerVersionInfo"
                    }
                },
                "type": "object"
            },
            "internal_server.ServerNamespaceInfo": {
                "properties": {
                    "all": {
                        "description": "All namespaces (if discoverable)",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "withSessions": {
                        "description": "Namespaces containing sessions",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
            "internal_server.ServerVersionInfo": {
                "properties": {
                    "buildDate": {
                        "type": "string"
                    },
                    "gitCommit": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_server.ServiceAccountInfo": {
                "properties": {
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.NamespacePermissions"
                    },
                    "session": {
                        "additionalProperties": {
                            "type": "boolean"
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "internal_server.SessionCreateRequest": {
                "description": "Request body for creating a new codespace session",
                "properties": {
                    "auth": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.AuthSpec"
                    },
                    "home": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.PVCSpec"
                    },
                    "name": {
                        "example": "my-session",
                        "type": "string"
                    },
                    "namespace": {
                        "example": "default",
                        "type": "string"
                    },
                    "networking": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.NetSpec"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.ProfileSpec"
                    },
                    "replicas": {
                        "example": 1,
                        "type": "integer"
                    },
                    "scratch": {
                        "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.PVCSpec"
                    }
                },
                "required": [
                    "name",
                    "profile"
                ],
                "type": "object"
            },
            "internal_server.SessionListResponse": {
                "description": "Response containing list of sessions with metadata",
                "properties": {
                    "filtered": {
                        "type": "boolean"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/github_com_codespace-operator_codespace-operator_api_v1.Session"
                        },
                        "type": "array"
                    },
                    "namespaces": {
                        "example": [
                            "default",
                            "kube-system"
                        ],
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "total": {
                        "example": 5,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "internal_server.SystemCapabilities": {
                "properties": {
                    "clusterScope": {
                        "description": "Displays cluster_scope: true or false",
                        "type": "boolean"
                    },
                    "multiTenant": {
                        "description": "Whether system supports multiple tenants",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    }
}